ieee/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
websockets==12.0
opencv-python-headless==4.10.0.84
ultralytics==8.2.0
numpy==1.26.4
scipy==1.13.0
supabase==2.6.0
python-dotenv==1.0.1
Pillow==10.4.0
aiofiles==24.1.0
